Fixed bug in sampling of angle. Changed Bank type to have uid as integer(8) to allow for more than 2^32 - 1 particles to be simulated.

+++ src/types.f90
@@ -74,23 +74,23 @@ module types
 !=====================================================================
 
   type Particle
-    integer :: uid          ! Unique ID
-    integer :: type         ! Particle type (n, p, e, etc)
-    real(8) :: xyz(3)       ! location
-    real(8) :: xyz_local(3) ! local location (after transformations)
-    real(8) :: uvw(3)       ! directional cosines
-    real(8) :: wgt          ! particle weight
-    real(8) :: E            ! energy
-    integer :: IE           ! index on energy grid
-    real(8) :: interp       ! interpolation factor for energy grid
-    integer :: cell         ! current cell
-    integer :: universe     ! current universe
-    integer :: lattice      ! current lattice
-    integer :: surface      ! current surface
-    integer :: index_x      ! lattice index for x direction
-    integer :: index_y      ! lattice index for y direction
-    logical :: alive        ! is particle alive?
-    integer :: n_coll       ! # of collisions
+    integer(8) :: uid          ! Unique ID
+    integer    :: type         ! Particle type (n, p, e, etc)
+    real(8)    :: xyz(3)       ! location
+    real(8)    :: xyz_local(3) ! local location (after transformations)
+    real(8)    :: uvw(3)       ! directional cosines
+    real(8)    :: wgt          ! particle weight
+    real(8)    :: E            ! energy
+    integer    :: IE           ! index on energy grid
+    real(8)    :: interp       ! interpolation factor for energy grid
+    integer    :: cell         ! current cell
+    integer    :: universe     ! current universe
+    integer    :: lattice      ! current lattice
+    integer    :: surface      ! current surface
+    integer    :: index_x      ! lattice index for x direction
+    integer    :: index_y      ! lattice index for y direction
+    logical    :: alive        ! is particle alive?
+    integer    :: n_coll       ! # of collisions
   end type Particle
 
 !=====================================================================
@@ -100,10 +100,10 @@ module types
 !=====================================================================
 
   type Bank
-     integer :: uid    ! Unique ID
-     real(8) :: xyz(3) ! location of bank particle
-     real(8) :: uvw(3) ! diretional cosines
-     real(8) :: E      ! energy
+     integer(8) :: uid    ! Unique ID
+     real(8)    :: xyz(3) ! location of bank particle
+     real(8)    :: uvw(3) ! diretional cosines
+     real(8)    :: E      ! energy
   end type Bank
 
 !=====================================================================
